requirements.txt
requests
apscheduler
python-telegram-bot==20.7
python-dotenv
pytz
numpy
fastapi
uvicorn
pycryptodome
pydantic
